WIP attempt to share dbus connection between objects
Expanded the dbus-macros-rs code into the file, then attempted to get
and share a connection.

Doesn't work as is because its trying to share between two objects.
Need to figure out how I can make the "Server" and "EntryGroup" part of
some kind of global object so that they can share the connection
successfully.

Haven't yet learnt enough rust to do that.. WIP :-)